Tests moved into e2e sub-folder
