+++ tests/e2e/extensions.cy.js
@@ -2,10 +2,10 @@
  * extensions.cy.js – Cypress test suite for ../src/extensions.js custom commands
  */
 
-import { registerCommands } from './support/registerCommands';
+import { registerCommands } from '../support/registerCommands';
 // Add attachFile() Cypress custom command, which is used in installExtensionFromFileUpload()
 import 'cypress-file-upload';
-import { config, caughtJavaScriptExceptions } from './support/setup';
+import { config, caughtJavaScriptExceptions } from '../support/setup';
 
 registerCommands();
