+++ tests/e2e/support.cy.js
@@ -2,8 +2,8 @@
  * support.cy.js – Cypress test suite for ../src/support.js custom commands
  */
 
-import { registerCommands } from './support/registerCommands';
-import { config, caughtJavaScriptExceptions } from './support/setup';
+import { registerCommands } from '../support/registerCommands';
+import { config, caughtJavaScriptExceptions } from '../support/setup';
 
 registerCommands();
